fix: Move HostPortPair type to common/configs
